Refactor code comment generation in CodeGlass extension
Removed verbose logging and improved process for generating commented code in the CodeGlass extension. The refactor also introduced temporary file handling in VSCode. The commented code is now directly edited in an open temporary document next to the original file, instead of saving the commented code to a new file.